+++ lib/commands/help.rb
@@ -53,6 +53,7 @@ module SlackSup
         set message [message]      - set the message users see when creating a S'Up DM
         unset message              - reset the message users see when creating a S'Up DM to the default one
         opt [in|out] [@mention]    - opt users in/out by @mention
+        promote|demote [@mention]  - promote or demote another user to/from channel admin
 
         S'Up
         ----